+++ modules/rest_modules/models.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class ChatRequest(BaseModel):
+    question: str
+    clear_history: bool = False
