+++ modules/rest_modules/rest_utils/image_utils/grainger_image_util.py
