+++ utils/deployment_scripts/docker-compose.yml
@@ -12,8 +12,8 @@ services:
 
 secrets:
   aws_access_key_id:
-    file: ../../aws_access_key_id
+    file: ../../secrets/aws_access_key_id
   aws_secret_access_key:
-    file: ../../aws_secret_access_key
+    file: ../../secrets/aws_secret_access_key
   bedrock_assume_role:
-    file: ../../bedrock_assume_role
+    file: ../../secrets/bedrock_assume_role
